+++ ml_core.lua
@@ -73,8 +73,8 @@ function RCLootCouncilML:GetItemInfo(item)
 			["subType"]		= subType,
 			["texture"]		= texture,
 			["boe"]			= bindType == LE_ITEM_BIND_ON_EQUIP,
-			["relic"]		= itemID and IsArtifactRelicItem(itemID) and select(3, C_ArtifactUI.GetRelicInfoByItemID(itemID)),
-			["token"]		= itemID and RCTokenTable[itemID],
+			--["relic"]		= itemID and IsArtifactRelicItem(itemID) and select(3, C_ArtifactUI.GetRelicInfoByItemID(itemID)),
+			--["token"]		= itemID and RCTokenTable[itemID],
 			["typeID"]		= typeID,
 			["subTypeID"]	= subTypeID,
 			["classes"]		= addon:GetItemClassesAllowedFlag(link)
@@ -424,70 +424,48 @@ end
 function RCLootCouncilML:UpdateMLdb()
 	-- The db has changed, so update the mldb and send the changes
 	addon:Debug("UpdateMLdb")
-	addon.mldb = self:BuildMLdb()
+	addon:OnMLDBReceived(self:BuildMLdb())
 	addon:SendCommand("group", "MLdb", addon.mldb)
 end
 
 function RCLootCouncilML:BuildMLdb()
-	-- Extract changes to normal responses/buttons
+	-- Extract changes to responses/buttons
 	local changedResponses = {};
-	for i = 1, db.numButtons do
-		if db.responses[i].text ~= addon.defaults.profile.responses[i].text or unpack(db.responses[i].color) ~= unpack(addon.defaults.profile.responses[i].color) then
-			changedResponses[i] = db.responses[i]
+	for type, responses in pairs(db.responses) do
+		for i in ipairs(responses) do
+			if i > db.buttons[type].numButtons then break end
+			if not addon.defaults.profile.responses[type]
+			or db.responses[type][i].text ~= addon.defaults.profile.responses[type][i].text
+			or unpack(db.responses[type][i].color) ~= unpack(addon.defaults.profile.responses[type][i].color) then
+				if not changedResponses[type] then changedResponses[type] = {} end
+				changedResponses[type][i] = db.responses[type][i]
+			end
 		end
 	end
 	local changedButtons = {};
-	for i = 1, db.numButtons do
-		if db.buttons[i].text ~= addon.defaults.profile.buttons[i].text then
-			changedButtons[i] = {text = db.buttons[i].text}
-		end
-	end
-	local changedTierButtons
-	if db.tierButtonsEnabled then
-		changedResponses.tier,changedTierButtons = {}, {}
-		for k,v in pairs(db.responses.tier) do
-			if v.text ~= addon.defaults.profile.responses.tier[k].text or unpack(v.color) ~= unpack(addon.defaults.profile.responses.tier[k].color) then
-				changedResponses.tier[k] = v
-			end
-		end
-		for i = 1, db.tierNumButtons do
-			if db.tierButtons[i].text ~= addon.defaults.profile.tierButtons[i].text then
-				changedTierButtons[i] = {text = db.tierButtons[i].text}
+	for type, buttons in pairs(db.buttons) do
+		for i in ipairs(buttons) do
+			if i > db.buttons[type].numButtons then break end
+			if not addon.defaults.profile.buttons[type]
+			or db.buttons[type][i].text ~= addon.defaults.profile.buttons[type][i].text then
+				if not changedButtons[type] then changedButtons[type] = {} end
+				changedButtons[type][i] = {text = db.buttons[type][i].text}
 			end
 		end
 	end
-	local changedRelicButtons
-	if db.relicButtonsEnabled then
-		changedResponses.relic, changedRelicButtons = {},{}
-		for k,v in pairs(db.responses.relic) do
-			if v.text ~= addon.defaults.profile.responses.relic[k].text or unpack(v.color) ~= unpack(addon.defaults.profile.responses.relic[k].color) then
-				changedResponses.relic[k] = v
-			end
-		end
-		for i = 1, db.relicNumButtons do
-			if db.relicButtons[i].text ~= addon.defaults.profile.relicButtons[i].text then
-				changedRelicButtons[i] = {text = db.relicButtons[i].text}
-			end
-		end
-	end
+	if changedButtons.default then changedButtons.default.numButtons = db.numButtons end
 
 	local MLdb = {
 		selfVote			= db.selfVote or nil,
 		multiVote		= db.multiVote or nil,
 		anonymousVoting= db.anonymousVoting or nil,
 		allowNotes		= db.allowNotes or nil,
-		numButtons		= db.numButtons,
-		tierNumButtons = db.tierNumButtons,
-		relicNumButtons= db.relicNumButtons,
+	--	numButtons		= db.numButtons,
 		hideVotes		= db.hideVotes or nil,
 		observe			= db.observe or nil,
 		buttons			= changedButtons,	-- REVIEW I'm not sure if it's feasible to nil out empty tables
-		tierButtons 	= changedTierButtons,
-		relicButtons 	= changedRelicButtons,
 		responses		= changedResponses,
 		timeout			= db.timeout,
-		tierButtonsEnabled = db.tierButtonsEnabled or nil,
-		relicButtonsEnabled = db.relicButtonsEnabled or nil,
 		rejectTrade 	= db.rejectTrade or nil
 	}
 
@@ -1283,7 +1261,7 @@ function RCLootCouncilML:TrackAndLogLoot(name, item, responseID, boss, votes, it
 	if reason and not reason.log then return end -- Reason says don't log
 	if not (db.sendHistory or db.enableHistory) then return end -- No reason to do stuff when we won't use it
 	if addon.testMode and not addon.nnp then return end -- We shouldn't track testing awards.
-	local _, link = GetItemInfo(item)
+	local _, link,_,_,_,_,_,_,equipLoc = GetItemInfo(item)
 	local i1,i2
 	if itemReplaced1 then i1 = select(2, GetItemInfo(itemReplaced1)) end
 	if itemReplaced2 then i2 = select(2, GetItemInfo(itemReplaced2)) end
@@ -1298,9 +1276,9 @@ function RCLootCouncilML:TrackAndLogLoot(name, item, responseID, boss, votes, it
 	history_table["votes"] 			= votes
 	history_table["itemReplaced1"]= i1
 	history_table["itemReplaced2"]= i2
-	history_table["response"] 		= reason and reason.text or addon:GetResponseText(responseID, tokenRoll, relicRoll)
+	history_table["response"] 		= reason and reason.text or addon:GetResponse(equipLoc,responseID).text
 	history_table["responseID"] 	= responseID or reason.sort - 400 															-- Changed in v2.0 (reason responseID was 0 pre v2.0)
-	history_table["color"]			= reason and reason.color or {addon:GetResponseColor(responseID, tokenRoll, relicRoll)}	-- New in v2.0
+	history_table["color"]			= reason and reason.color or addon:GetResponse(equipLoc, responseID).color	-- New in v2.0
 	history_table["class"]			= self.candidates[name].class																-- New in v2.0
 	history_table["isAwardReason"]= reason and true or false																	-- New in v2.0
 	history_table["difficultyID"]	= difficultyID																					-- New in v2.3+